+++ Changes.md
@@ -1,7 +1,10 @@
 1.3.x.x (relative to 1.3.14.0)
 =======
 
+Features
+--------
 
+- RenderPassEditor : Added "Camera Inclusions" and "Camera Exclusions" columns, providing control over the camera visibility of scene locations in each render pass.
 
 1.3.14.0 (relative to 1.3.13.1)
 ========